+++ package.json
@@ -38,7 +38,7 @@
     "test:unit": "vitest --environment jsdom --root src/",
     "test:unit:update-snapshots": "npm run test:unit -- -u",
     "build:vite": "rimraf dist && vite build",
-    "build:css": "rimraf dist_css && mkdir -p dist_css/tailwind && node ./scripts/build-tailwind-css-files.cjs",
+    "build:css": "rimraf dist_css && mkdir -p dist_css/tailwind && node ./scripts/build-tailwind-css-files.js",
     "build:tailwind": "node ./src/tokens/scripts/convertToTailwind.cjs",
     "type-check": "vue-tsc --noEmit -p tsconfig.vitest.json --composite false",
     "build:types": "rimraf dist_types && vue-tsc --declaration --emitDeclarationOnly -p tsconfig.compile.json",
